test(extract/nvd/feed/cve/v2): fold helper tests into v2_test via export_test
Move the cpematch range-helper unit tests into the external v2_test package
(next to TestExtract) and expose the unexported helpers — parseRange,
versionInBounds, unescapeWFN — through export_test.go. The unexported
semverBounds is mirrored as an exported Bounds so the external test can build
and inspect bounds with named fields. Removes v2_internal_test.go.
